Refactor config
